Move splitter SplitterError from GangaCMS.Lib.Utils to Ganga.Core
