--- python/GangaCMS/Lib/Utils/__init__.py
@@ -1,1 +1,0 @@
-from SplitterError import *
